page_queue is constrained by a size supplied in options, default 100.
This alleviates a problem we experienced with very responsive sites (wikipedia)
and a moderate per-page processing time. The page_queue would grow much faster
than it could be drained, using more and more RAM.

This change means that the queue grows until full, at which point the
Tentacles will block until the queue shrinks.

This does have the impact of slowing the crawl in some cases.